// front/src/app/api-client/model/reminder-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FamilyMemberDTO } from './family-member-dto';
import { RecurrenceRuleDTO } from './recurrence-rule-dto';


/**
 * Reminder
 */
export interface ReminderDTO { 
    id?: number;
    titre?: string;
    description?: string;
    type?: ReminderDTO.TypeEnum;
    dateHeureDebut?: string;
    dateHeureFin?: string;
    membres?: Array<FamilyMemberDTO>;
    isRecurring?: boolean;
    /**
     * Présent uniquement si isRecurring = true
     */
    recurrence?: RecurrenceRuleDTO;
}
export namespace ReminderDTO {
    export const TypeEnum = {
        Sport: 'SPORT',
        TestMedical: 'TEST_MEDICAL',
        RendezVous: 'RENDEZ_VOUS',
        Autre: 'AUTRE',
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}